LPS-61205 Move components to wiki-web to be consistent with the rest of the modules
